+++ common/src/main/kotlin/com/cobblemon/mod/common/CobblemonCommands.kt
@@ -63,6 +63,7 @@ object CobblemonCommands {
         SpectateBattleCommand.register(dispatcher)
         CobblemonConfigCommand.register(dispatcher)
         UnlockPCBoxWallpaperCommand.register(dispatcher)
+        ChangePCBoxesCommand.register(dispatcher)
 
         // Possibly lock down registration if and only if under dev environment or running in an environment
         // with a certain system environment variable set